update: fix miller indices

+++ src/py/mat3ra/esse/models/material/reusable/slab/miller_indices.py
@@ -9,8 +9,8 @@ from typing import List, Optional
 from pydantic import Field, RootModel
 
 
-class MillerIndicesSchema(RootModel[Optional[List[float]]]):
-    root: Optional[List[float]] = Field([0, 0, 1], max_length=3, min_length=3, title="Miller Indices Schema")
+class MillerIndicesSchema(RootModel[Optional[List[int]]]):
+    root: Optional[List[int]] = Field([0, 0, 1], max_length=3, min_length=3, title="Miller Indices Schema")
     """
     Miller indices for crystallographic plane designation
     """
